fix: use kimi-k2 for signal generation
