+++ package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "@langchain/community": "^0.3.49",
     "@langchain/core": "^0.3.66",
+    "@langchain/groq": "^0.2.3",
     "@langchain/langgraph": "^0.3.11",
     "@langchain/mcp-adapters": "^0.6.0",
     "@langchain/openai": "^0.5.18",
