+++ src/agents/model.ts
@@ -1,3 +1,4 @@
+import { ChatGroq } from "@langchain/groq";
 import { ChatOpenAI } from "@langchain/openai";
 
 export const gpt4o = new ChatOpenAI({
@@ -10,4 +11,14 @@ export const gpt4oMini = new ChatOpenAI({
   modelName: "gpt-4o-mini",
   temperature: 0.7,
   apiKey: process.env.OPENAI_API_KEY,
+});
+
+export const kimiK2 = new ChatGroq({
+  apiKey: process.env.GROQ_API_KEY,
+  model: "moonshotai/kimi-k2-instruct",
+});
+
+export const llama3370bVersatile = new ChatGroq({
+  apiKey: process.env.GROQ_API_KEY,
+  model: "llama-3.3-70b-versatile",
 });
